pyproject.toml
[tool.poetry]
name = "matias_django_blog"
version = "0.1.0"
description = "Matias personal blog & more"
authors = ["Matias Cardenas <cardenasmatias.1990@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.1"
black = "^22.8.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
django-tinymce = "^3.5.0"
psycopg2 = "^2.9.5"
djangorestframework = "^3.14.0"
pyyaml = "^6.0"
uritemplate = "^4.1.1"
drf-openapi-tester = "^2.1.3"
pytest = "^7.2.0"
pytest-django = "^4.5.2"
freezegun = "^1.2.2"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"
mistune = "^3.0.2"
pygments = "^2.17.2"
djlint = "^1.34.1"
pre-commit = "^3.6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.black]
line-length = 120

[tool.djlint]
profile="django"